Write leg2 summary (#369)
* Write summary output for leg2 of trade demand calculation, move functions that write summaries to freight_utils
